Clip eBird range map
Clipped eBird range map to reduce output file size. Only eBird range
that is not printed to output file is removed. Size of clipped area was
selected by experimentation and is only approximate. The size of the
clipping bounding box may need adjusting if size of printed area or
printing device is changed.

Changed reading of eBird range from readORG to st_read because
st_intersection expects an sf object.